made some formatting/renaming changes
