+++ lib/drivers/carelink/cli/csv_loader.js
@@ -4,42 +4,49 @@ var program = require('commander');
 var fs = require('fs');
 
 var driverManager = require('../../../driverManager');
-var pwdSimulator = require('../carelinkSimulator.js');
-var carelinkDriver = require('../carelinkDriver');
+var libreViewDriver = require('../libreViewDriver');
 var api = require('../../../core/api.js');
 var config = require('../../../../.config.js');
 var pkg = require('../../../../package.json');
 var builder = require('../../../objectBuilder');
-var TZOUtil = require('../../../TimezoneOffsetUtil');
 
-var intro = 'CareLink CLI:';
-
+var intro = 'LibreView CLI:';
 
 /*
- * Load the given carelink file and then parse and send the data to the tp-platform
+ * Load the given LibreView file and then parse and send the data to the tp-platform
  */
 function loadFile(filePath, tz, userid, dev_mode) {
     console.log(intro, 'Reading', filePath);
     fs.readFile(filePath, 'utf8', function(error, data) {
         if (error) {
-            console.log(intro, 'Error reading CareLink file', error);
+            console.log(intro, 'Error reading LibreView file', error);
             return;
         }
 
-        var drivers = {'Carelink': carelinkDriver(pwdSimulator, api)};
-        var cfg = { 'Carelink': { filename: filePath, fileData: data, timezone: tz, version: pkg.version, groupId: userid, builder: builder(), api: api, dev_mode: dev_mode} };
+        var drivers = {'LibreView': libreViewDriver};
+        var cfg = { 'LibreView':
+          {
+            filename: filePath,
+            fileData: data,
+            timezone: tz,
+            version: pkg.version,
+            groupId: userid,
+            builder: builder(),
+            api: api
+          }
+        };
 
-        processCarelink(driverManager(drivers, cfg));
+        processLibreView(driverManager(drivers, cfg));
     });
 }
 
 /*
- * Process our raw carelink data
+ * Process our raw LibreView data
  */
-function processCarelink(driverMgr){
-    driverMgr.process('Carelink', function(err, result) {
+function processLibreView(driverMgr){
+    driverMgr.process('LibreView', function(err, result) {
         if (err) {
-            console.log(intro, 'Error processing CareLink data:', err);
+            console.log(intro, 'Error processing LibreView data:', err);
             process.exit();
         }
 
@@ -56,7 +63,7 @@ function login(un, pw, config, cb){
     apiUrl: config.API_URL,
     uploadUrl: config.UPLOAD_URL,
     dataUrl: config.DATA_URL,
-    version: 'uploader node CLI tool - CareLink',
+    version: 'uploader node CLI tool - LibreView',
     builder: builder(),
   });
   api.init(function() {
@@ -69,31 +76,24 @@ function login(un, pw, config, cb){
 
 program
     .version('0.0.1')
-    .option('-f, --file [path]', 'carelink csv file path')
+    .option('-f, --file [path]', 'LibreView csv file path')
     .option('-u, --username [user]', 'username')
     .option('-p, --password [pw]', 'password')
-    .option('-i, --id [id]', 'user id') /* dev only */
-    .option('-t, --timezone [tz]', 'named timezone', config.DEFAULT_TIMEZONE)//default is 'America/Los_Angeles'
+    .option('-t, --timezone [tz]', 'named timezone', config.DEFAULT_TIMEZONE) // default is 'America/Los_Angeles'
     .parse(process.argv);
 
 function load(program) {
-    console.log(intro, 'Loading CareLink CSV...');
+    console.log(intro, 'Loading LibreView CSV...');
 
     if(!(program.file && program.username && program.password)) {
         program.help();
-        return
+        return;
     }
     if (!fs.existsSync(program.file)){
-        console.log(intro + ' CSV file at [%s] not found', program.carelink);
-        return
+        console.log(intro + ' CSV file at [%s] not found', program.LibreView);
+        return;
     }
 
-    var userid = program.id;
-    if (userid !== true) { /* dev only */
-        console.log(intro, "DEV MODE", userid)
-        loadFile(program.file, program.timezone, userid, true);
-        return
-    }
     login(program.username, program.password, config, function(err, data){
         if (err) {
             console.log(intro, 'Failed authentication!');
@@ -108,4 +108,4 @@ function load(program) {
     });
 }
 
-load(program)
+load(program);
